# docs/swagger.yaml
basePath: /swagger/visit-service
definitions:
  common.Page:
    properties:
      items: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  common.Response:
    properties:
      data: {}
      msg:
        type: string
      status:
        type: integer
    type: object
...
  model.Camera:
    properties:
      device_name:
        description: 设备名称
        type: string
      device_no:
        description: 设备编号
        type: string
      device_type:
        description: 设备类型
        type: integer
      id:
        description: 摄像头ID
        type: string
      location_type:
        description: 位置类型
        type: integer
      main_stream_url:
        description: 主码流URL
        type: string
      manufacturer:
        description: 设备厂商
        type: string
      rel_vr_camera_id:
        description: 关联VR摄像头ID(当位置为可移动时，需关联VR摄像头)
        type: string
      status:
        description: 状态(0:正常,1:禁用)
        type: integer
      sub_stream_url:
        description: 辅码流URL
        type: string
    type: object
...
info:
  contact: {}
  description: visit-service API
  title: visit-service API
  version: "1.0"
paths:
...
  /camera:
    get:
      description: query objects
      parameters:
      - description: _select
        in: query
        name: _select
        type: string
      - description: order
        in: query
        name: _order
        type: string
      - description: id
        in: query
        name: id
        type: string
      - description: device_name
        in: query
        name: device_name
        type: string
      - description: device_no
        in: query
        name: device_no
        type: string
      - description: location_type
        in: query
        name: location_type
        type: string
      - description: device_type
        in: query
        name: device_type
        type: string
      - description: manufacturer
        in: query
        name: manufacturer
        type: string
      - description: main_stream_url
        in: query
        name: main_stream_url
        type: string
      - description: sub_stream_url
        in: query
        name: sub_stream_url
        type: string
      - description: rel_vr_camera_id
        in: query
        name: rel_vr_camera_id
        type: string
      - description: status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: objects array
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Camera'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: query objects
      tags:
      - 摄像头
    post:
      consumes:
      - application/json
      description: save
      parameters:
      - description: object
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/model.Camera'
      produces:
      - application/json
      responses:
        "200":
          description: object
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Camera'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: save
      tags:
      - 摄像头
...
  /camera/{id}:
    delete:
      description: delete
      parameters:
      - description: 实例id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: object
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Camera'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: delete
      tags:
      - 摄像头
  /camera/batch-delete:
    post:
      consumes:
      - application/json
      description: batch delete
      parameters:
      - description: id array
        in: body
        name: ids
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: batch delete
      tags:
      - 摄像头
  /camera/batch-upsert:
    post:
      consumes:
      - application/json
      description: batch update
      parameters:
      - description: objects array
        in: body
        name: entities
        required: true
        schema:
          items:
            additionalProperties: true
            type: object
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: batch update
      tags:
      - 摄像头
...
  /camera/page:
    get:
      description: page query, _page(from 1 begin), _page_size, _order, and others
        fields, status=1, name=$like.%CAM%
      parameters:
      - description: current page
        in: query
        name: _page
        required: true
        type: integer
      - description: page size
        in: query
        name: _page_size
        required: true
        type: integer
      - description: order
        in: query
        name: _order
        type: string
      - description: id
        in: query
        name: id
        type: string
      - description: device_name
        in: query
        name: device_name
        type: string
      - description: device_no
        in: query
        name: device_no
        type: string
      - description: location_type
        in: query
        name: location_type
        type: string
      - description: device_type
        in: query
        name: device_type
        type: string
      - description: manufacturer
        in: query
        name: manufacturer
        type: string
      - description: main_stream_url
        in: query
        name: main_stream_url
        type: string
      - description: sub_stream_url
        in: query
        name: sub_stream_url
        type: string
      - description: rel_vr_camera_id
        in: query
        name: rel_vr_camera_id
        type: string
      - description: status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: objects array
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/common.Page'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/model.Camera'
                        type: array
                    type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: page query
      tags:
      - 摄像头
...
  /manual/delete-visit-schedule:
    post:
      consumes:
      - application/json
      description: 手动删除排班
      parameters:
      - description: 删除开始时间，2006-01-02
        in: query
        name: start_day
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.Response'
      summary: 手动删除排班
      tags:
      - Manual
  /manual/init-visit-schedule:
    post:
      consumes:
      - application/json
      description: 手动初始化排班
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.Response'
      summary: 手动初始化排班
      tags:
      - Manual
...
swagger: "2.0"